- speedup for fog of war algorithm (about 10 % less total CPU usage on Android)

+++ src/jsettlers/logic/algorithms/fogofwar/IFogOfWarGrid.java
@@ -11,8 +11,8 @@ import jsettlers.common.movable.IMovable;
  */
 public interface IFogOfWarGrid {
 
-	IMovable getMovableAt(int x, int y);
+	IMovable getMovableAt(short x, short y);
 
-	IMapObject getMapObjectsAt(int x, int y);
+	IMapObject getMapObjectsAt(short x, short y);
 
 }
